systemhotelarski_backoffice/requirements.txt
Django==1.8.5
djangorestframework==3.2.3